refactor(mejoras): split monolithic component into modular architecture
