+++ src/features/mejoras/presentation/components/mejora-search-results/README.md
@@ -1,0 +1,188 @@
+# Mejora Search Results Components
+
+## 📁 Estructura
+
+Esta carpeta contiene los componentes refactorizados para mostrar los resultados de búsqueda de mejoras, siguiendo principios de **Clean Code** y **Single Responsibility Principle**.
+
+```
+mejora-search-results/
+├── constants.ts                          # Constantes compartidas (DEFAULT_APPROVER_ID)
+├── formatters.ts                         # Funciones de formateo (fechas, nombres, ubicaciones)
+├── mejora-detail-dialog.component.tsx    # Diálogo modal con detalles de una mejora
+├── related-mejoras-panel.component.tsx   # Panel de mejoras relacionadas con acciones
+├── fallecidos-results.component.tsx      # Vistas para resultados de fallecidos
+├── propietarios-results.component.tsx    # Vista para resultados de propietarios
+├── index.ts                              # Barrel export
+└── README.md                             # Esta documentación
+```
+
+## 🎯 Propósito
+
+Antes teníamos **829 líneas** en un solo archivo. Ahora está dividido en **módulos especializados** de ~50-200 líneas cada uno.
+
+### Beneficios de la Refactorización:
+
+✅ **Single Responsibility**: Cada archivo tiene una única responsabilidad  
+✅ **Reusabilidad**: Los componentes pueden importarse independientemente  
+✅ **Mantenibilidad**: Más fácil encontrar y modificar código específico  
+✅ **Testabilidad**: Cada componente puede testearse de forma aislada  
+✅ **Legibilidad**: Código más claro y fácil de entender  
+
+## 📦 Componentes
+
+### `constants.ts`
+Constantes compartidas entre componentes.
+
+```typescript
+export const DEFAULT_APPROVER_ID = "...";
+```
+
+### `formatters.ts`
+Funciones puras para formatear datos.
+
+```typescript
+export const formatDate = (value?: string): string
+export const fullName = (person?: { nombres?: string; apellidos?: string }): string
+export const formatNichoLocation = (nicho?: {...}): string
+```
+
+### `mejora-detail-dialog.component.tsx`
+Diálogo modal que muestra todos los detalles de una mejora.
+
+**Props:**
+- `mejora: MejoraEntity`
+
+**Características:**
+- Vista completa de información general, solicitante, fallecido, programación, nicho
+- Diseño responsive con scroll vertical
+- Usa Dialog de shadcn/ui
+
+### `related-mejoras-panel.component.tsx`
+Panel que muestra mejoras relacionadas con acciones (aprobar, editar, descargar).
+
+**Props:**
+- `mejoras: MejoraEntity[]`
+- `isLoading?: boolean`
+- `searchTerm: string`
+
+**Características:**
+- Tabla con información resumida de mejoras
+- Botones de acción según estado (Solicitado/Aprobado)
+- Integración con mutations (approve, download)
+- Loading states
+
+### `fallecidos-results.component.tsx`
+Dos componentes para mostrar resultados de fallecidos.
+
+**Componentes:**
+1. **SingleResultView**: Vista detallada para un solo resultado
+2. **MultipleResultsView**: Lista de tarjetas para múltiples resultados
+
+**Props compartidos:**
+- `searchTerm: string`
+
+**Características:**
+- Muestra información del fallecido y solicitante
+- Lista de requisitos de inhumación asociados
+- Botón para crear mejora desde cada requisito
+
+### `propietarios-results.component.tsx`
+Vista para mostrar propietarios de nichos encontrados.
+
+**Props:**
+- `propietarios: MejoraSearchAllResultsEntity["propietarios"]`
+- `searchTerm: string`
+
+**Características:**
+- Tarjetas por propietario con sus nichos
+- Tabla con información detallada de cada nicho
+- Filtrado visual de nichos activos/inactivos
+- Botón para crear mejora solo en nichos activos
+
+## 🔄 Flujo de Datos
+
+```
+mejora-search-results.component.tsx (Orquestador)
+    ↓
+    ├─→ SingleResultView / MultipleResultsView (Fallecidos)
+    ├─→ PropietariosResultsView (Propietarios)
+    └─→ RelatedMejorasPanel (Mejoras relacionadas)
+            ↓
+            └─→ MejoraDetailDialog (Modal de detalles)
+```
+
+## 📝 Uso
+
+### Importación desde el componente principal
+
+```typescript
+import MejoraSearchResults from "@/features/mejoras/presentation/components/mejora-search-results.component";
+
+<MejoraSearchResults
+  results={searchResults}
+  searchTerm={query}
+  relatedMejoras={mejoras}
+  isLoadingRelated={isLoading}
+/>
+```
+
+### Importación de subcomponentes (si es necesario)
+
+```typescript
+import { 
+  MejoraDetailDialog, 
+  formatDate, 
+  fullName 
+} from "@/features/mejoras/presentation/components/mejora-search-results";
+```
+
+## 🧪 Testing
+
+Cada componente puede testearse independientemente:
+
+```typescript
+// Ejemplo de test para formatters
+describe('formatDate', () => {
+  it('should format valid date', () => {
+    expect(formatDate('2024-01-15')).toBe('15/1/2024');
+  });
+});
+```
+
+## 🔧 Mantenimiento
+
+### Para agregar una nueva funcionalidad:
+
+1. **Formateo de datos**: Agregar en `formatters.ts`
+2. **Constantes**: Agregar en `constants.ts`
+3. **Nuevo componente**: Crear archivo separado en la carpeta
+4. **Exportar**: Actualizar `index.ts`
+
+### Para modificar un componente existente:
+
+1. Localizar el archivo específico (nombre descriptivo)
+2. Modificar solo ese componente
+3. Los cambios no afectan otros componentes
+
+## 📊 Métricas de la Refactorización
+
+| Métrica | Antes | Después |
+|---------|-------|---------|
+| Líneas totales | 829 | ~829 (distribuidas) |
+| Archivos | 1 | 7 |
+| Líneas por archivo | 829 | 50-250 |
+| Responsabilidades | Múltiples | Una por archivo |
+| Reusabilidad | Baja | Alta |
+| Testabilidad | Difícil | Fácil |
+
+## 🚀 Futuras Mejoras
+
+- [ ] Agregar tests unitarios para cada componente
+- [ ] Crear Storybook stories para cada componente
+- [ ] Memoizar componentes con React.memo si hay problemas de performance
+- [ ] Extraer tipos compartidos a un archivo `types.ts`
+
+---
+
+**Última actualización**: Noviembre 2025  
+**Autor**: Refactorización Clean Code
