+++ src/features/mejoras/presentation/components/mejora-search-results/constants.ts
@@ -1,0 +1,5 @@
+/**
+ * Constants for mejora search results
+ */
+
+export const DEFAULT_APPROVER_ID = "11657f06-85d6-42bb-84f6-7e3ffe06965d";
