+++ src/features/mejoras/presentation/components/mejora-search-results/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Barrel export for mejora-search-results components
+ * Provides a clean interface for importing components and utilities
+ */
+
+export { MejoraDetailDialog } from "./mejora-detail-dialog.component";
+export { RelatedMejorasPanel } from "./related-mejoras-panel.component";
+export { SingleResultView, MultipleResultsView } from "./fallecidos-results.component";
+export { PropietariosResultsView } from "./propietarios-results.component";
+export { formatDate, fullName, formatNichoLocation } from "./formatters";
+export { DEFAULT_APPROVER_ID } from "./constants";
